Split join and goto slash commands, the latter shouldn't auto_join
